Add unit tests, LVL dump tool, and GitHub Pages deployment
- 70 unit tests (CRC hashing, chunk reader, ODF parser, camera math)
- lvl_dump CLI tool for inspecting .lvl file chunk structure
- GitHub Pages deployment on gh-pages branch

+++ app/openswbf/main.cpp
@@ -181,8 +181,8 @@ bool init_terrain() {
 
     for (int row = 0; row < GRID; ++row) {
         for (int col = 0; col < GRID; ++col) {
-            float wx = col * SCALE - HALF;
-            float wz = row * SCALE - HALF;
+            float wx = static_cast<float>(col) * SCALE - HALF;
+            float wz = static_cast<float>(row) * SCALE - HALF;
             float wy = terrain_height(wx, wz);
 
             // Compute normal via central differences.
